refactor: refactored Artist Details.
