+++ ui/src/main/java/com/cairosquad/ui/details/artist/ArtistScreen.kt
@@ -1,62 +1,41 @@
 package com.cairosquad.ui.details.artist
 
-import android.os.Build
 import android.widget.Toast
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.navigationBars
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.statusBars
 import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.text.BasicText
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.BlurredEdgeTreatment
-import androidx.compose.ui.draw.blur
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Brush.Companion.verticalGradient
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.lerp
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
-import com.cairosquad.design_system.R
 import com.cairosquad.design_system.basic_component.AppBar
-import com.cairosquad.design_system.basic_component.ExpandableText
-import com.cairosquad.design_system.basic_component.Icon
-import com.cairosquad.design_system.basic_component.InfoChip
-import com.cairosquad.design_system.modifier.CustomBrush
 import com.cairosquad.design_system.theme.Theme
-import com.cairosquad.safe_image_viewer.safe_image_viewer.SafeImageViewer
-import com.cairosquad.ui.BuildConfig
 import com.cairosquad.ui.details.DetailsFailContent
-import com.cairosquad.ui.movio_component.LoadingMovieCard
-import com.cairosquad.ui.movio_component.LoadingMovieImage
-import com.cairosquad.ui.movio_component.MovieCard
+import com.cairosquad.ui.details.artist.content.ArtistBackgroundImage
+import com.cairosquad.ui.details.artist.content.ArtistBiography
+import com.cairosquad.ui.details.artist.content.ArtistDepartment
+import com.cairosquad.ui.details.artist.content.ArtistImage
+import com.cairosquad.ui.details.artist.content.ArtistInformation
+import com.cairosquad.ui.details.artist.content.ArtistMoviesAndSeries
+import com.cairosquad.ui.details.artist.content.ArtistName
 import com.cairosquad.ui.navigation.MovieRoute
 import com.cairosquad.ui.navigation.SeriesRoute
 import com.cairosquad.ui.utils.ObserveAsEffect
@@ -65,10 +44,6 @@ import com.cairosquad.viewmodel.details.artist.ArtistEffect
 import com.cairosquad.viewmodel.details.artist.ArtistInteractionListener
 import com.cairosquad.viewmodel.details.artist.ArtistScreenState
 import com.cairosquad.viewmodel.details.artist.ArtistViewModel
-import java.text.SimpleDateFormat
-import java.util.Date
-import java.util.Locale
-
 
 @Composable
 fun ArtistScreen(
@@ -82,6 +57,17 @@ fun ArtistScreen(
         }
 
     val state by artistViewModel.screenState.collectAsState()
+
+    ArtistScreenEffects(artistViewModel, navController)
+
+    ArtistScreenContent(modifier = modifier, state = state, listener = artistViewModel)
+}
+
+@Composable
+private fun ArtistScreenEffects(
+    artistViewModel: ArtistViewModel,
+    navController: NavController
+) {
     val context = LocalContext.current
     ObserveAsEffect(artistViewModel.effect) { effect ->
         when (effect) {
@@ -106,8 +92,6 @@ fun ArtistScreen(
             }
         }
     }
-
-    ArtistScreenContent(modifier = modifier, state = state, listener = artistViewModel)
 }
 
 @Composable
@@ -125,10 +109,10 @@ private fun ArtistScreenContent(
 		lerp(Theme.color.surfaces.statusBarShadow, Theme.color.surfaces.surface, progress)
 	val animatedEndColor = lerp(Color.Transparent, Theme.color.surfaces.surface, progress)
 	val animatedBrush = verticalGradient(colors = listOf(animatedStartColor, animatedEndColor))
-	Box(modifier = modifier.fillMaxSize()) {
+
+    Box(modifier = modifier.fillMaxSize()) {
 		if (state.screenStatus == ArtistScreenState.ScreenStatus.FAILED) {
             DetailsFailContent(onTryAgainClick = listener::onRefresh)
-
         }
 		else{
 			Column(
@@ -144,286 +128,14 @@ private fun ArtistScreenContent(
 						.fillMaxWidth()
 						.height(340.dp)
 				) {
-					if (state.artist.photoPath.isBlank()) {
-						Box(
-							modifier = Modifier
-								.fillMaxSize()
-								.height(335.dp)
-								.offset(y = (-5).dp)
-								.CustomBrush(0.5f, 16.dp),
-						)
-					} else {
-						Box {
-							SafeImageViewer(
-								model = BuildConfig.IMAGE_BASE_URL + state.artist.photoPath,
-								contentDescription = "blured image",
-								modifier = Modifier
-									.fillMaxSize()
-									.height(335.dp)
-									.then(
-										if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-											Modifier.blur(16.dp, edgeTreatment = BlurredEdgeTreatment.Unbounded)
-										} else {
-											Modifier
-										}
-									)
-									.offset(y = (-20).dp),
-								blur = if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S) 16 else 0,
-								isBlurForced = true
-							)
-							if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S) {
-								Box(
-									modifier = Modifier
-										.fillMaxWidth()
-										.height(40.dp)
-										.align(Alignment.BottomCenter)
-										.background(
-											brush = verticalGradient(
-												colors = listOf(
-													Theme.color.surfaces.surface.copy(alpha = 0.00f),
-													Theme.color.surfaces.surface.copy(alpha = 0.10f),
-													Theme.color.surfaces.surface.copy(alpha = 0.50f),
-													Theme.color.surfaces.surface.copy(alpha = 0.90f),
-													Theme.color.surfaces.surface,
-												)
-											)
-										)
-								)
-							}
-
-						}
-					}
-					when (state.screenStatus) {
-						ArtistScreenState.ScreenStatus.LOADING -> {
-							LoadingMovieImage(
-								Modifier
-									.padding(top = 31.dp)
-									.clip(CircleShape)
-									.size(160.dp)
-							)
-						}
-
-                        ArtistScreenState.ScreenStatus.SUCCESS -> {
-                            if (state.artist.photoPath.isNotEmpty()) {
-                                SafeImageViewer(
-                                    model = BuildConfig.IMAGE_BASE_URL + state.artist.photoPath,
-                                    modifier = Modifier
-                                        .padding(horizontal = 6.67.dp)
-                                        .padding(top = 31.dp)
-                                        .size(160.dp)
-                                        .clip(CircleShape),
-                                    contentDescription = stringResource(R.string.artist_image),
-                                    nudeThreshold = 0.0,
-                                    nonNudeThreshold = 0.0,
-                                    loadingPlaceholder = {
-                                        LoadingMovieImage(
-                                            Modifier
-                                                .clip(CircleShape)
-                                                .size(160.dp)
-                                        )
-                                    }
-                                )
-                            } else {
-                                Box(
-                                    modifier = Modifier
-                                        .padding(horizontal = 6.67.dp)
-                                        .padding(top = 31.dp)
-                                        .size(160.dp)
-                                        .clip(CircleShape)
-                                        .background(Theme.color.system.defaultImageBackground),
-                                    contentAlignment = Alignment.Center
-                                ) {
-                                    Icon(
-                                        modifier = Modifier.size(24.dp),
-                                        imageVector = ImageVector.vectorResource(id = R.drawable.image_icon),
-                                        contentDescription = stringResource(R.string.default_image_icon),
-                                        tint = Color(0xFFEFF1F5)
-                                    )
-                                }
-                            }
-                        }
-
-                        ArtistScreenState.ScreenStatus.FAILED -> {}
-                    }
+					ArtistBackgroundImage(state = state)
+                    ArtistImage(state)
                 }
-                when (state.screenStatus) {
-                    ArtistScreenState.ScreenStatus.LOADING -> {
-                        LoadingMovieImage(
-                            Modifier
-                                .padding(horizontal = 16.dp)
-                                .clip(CircleShape)
-                                .size(width = 60.dp, height = 32.dp)
-                        )
-                    }
-
-                    ArtistScreenState.ScreenStatus.SUCCESS -> {
-                        BasicText(
-                            modifier = Modifier.padding(horizontal = 16.dp),
-                            text = state.artist.name,
-                            style = Theme.textStyle.headline.mediumMedium18
-                                .copy(color = Theme.color.surfaces.onSurface),
-                        )
-                    }
-
-                    ArtistScreenState.ScreenStatus.FAILED -> {}
-                }
-                when (state.screenStatus) {
-                    ArtistScreenState.ScreenStatus.LOADING -> {
-                        LoadingMovieImage(
-                            Modifier
-                                .padding(start = 16.dp, top = 4.dp)
-                                .clip(CircleShape)
-                                .size(width = 60.dp, height = 32.dp)
-                        )
-                    }
-
-                    ArtistScreenState.ScreenStatus.SUCCESS -> {
-                        BasicText(
-                            modifier = Modifier.padding(start = 16.dp, top = 4.dp),
-                            text = state.artist.department,
-                            style = Theme.textStyle.label.smallRegular14
-                                .copy(color = Theme.color.surfaces.onSurfaceVariant),
-                        )
-                    }
-
-                    ArtistScreenState.ScreenStatus.FAILED -> {}
-                }
-                when (state.screenStatus) {
-                    ArtistScreenState.ScreenStatus.LOADING -> {
-                        Row(
-                            modifier = Modifier.padding(horizontal = 16.dp),
-                            horizontalArrangement = Arrangement.spacedBy(8.dp)
-                        ) {
-                            repeat(2) {
-                                LoadingMovieImage(
-                                    Modifier
-                                        .padding(top = 16.dp, start = 16.dp)
-                                        .clip(CircleShape)
-                                        .size(width = 60.dp, height = 32.dp)
-                                )
-                            }
-                        }
-                    }
-
-					ArtistScreenState.ScreenStatus.SUCCESS -> {
-						LazyRow(
-							horizontalArrangement = Arrangement.spacedBy(8.dp),
-							modifier = Modifier.padding(top = 16.dp),
-							contentPadding = PaddingValues(horizontal = 16.dp)
-						) {
-							val lastWord = state.artist.country
-								.trim()
-								.let { full ->
-									val partsByComma = full.split(",")
-									val lastPart = partsByComma.lastOrNull()?.trim().orEmpty()
-									lastPart.split(" ").lastOrNull()?.trim()
-								}
-								?.takeIf { it.isNotBlank() }
-							state.artist.birthDate?.let{ birthDate ->
-								item {
-									InfoChip(
-										text = formatBirthDateLegacy(birthDate),
-										imgRes = R.drawable.date,
-									)
-								}
-							}
-							if (lastWord != null) {
-								item {
-									InfoChip(
-										text = state.artist.country,
-										imgRes = R.drawable.component_1,
-									)
-								}
-							}
-						}
-					}
-
-                    ArtistScreenState.ScreenStatus.FAILED -> {}
-                }
-                when (state.screenStatus) {
-                    ArtistScreenState.ScreenStatus.LOADING -> {
-                        LoadingMovieImage(
-                            Modifier
-                                .padding(16.dp)
-                                .height(120.dp)
-                                .fillMaxWidth()
-                        )
-                    }
-
-                    ArtistScreenState.ScreenStatus.SUCCESS -> {
-                        if (state.artist.biography.isNotBlank()) {
-                            ExpandableText(
-                                text = state.artist.biography,
-                                color = Theme.color.surfaces.onSurface,
-                                style = Theme.textStyle.label.smallRegular14,
-                                modifier = Modifier
-                                    .padding(16.dp)
-                                    .fillMaxWidth(),
-                                collapsedMaxLine = 5,
-                                showMoreText = stringResource(com.cairosquad.ui.R.string.read_more_with_dotes_behind),
-                                showMoreColor = Theme.color.surfaces.onSurfaceVariant,
-                                showLessText = stringResource(com.cairosquad.ui.R.string.read_less_with_dotes_behind)
-                            )
-                        }
-                    }
-
-                    ArtistScreenState.ScreenStatus.FAILED -> {}
-                }
-                when (state.screenStatus) {
-                    ArtistScreenState.ScreenStatus.LOADING -> {
-                        LoadingMovieImage(
-                            Modifier.size(height = 32.dp, width = 64.dp)
-                        )
-                        LazyRow {
-                            items(10) {
-                                LoadingMovieCard()
-                            }
-                        }
-                    }
-
-					ArtistScreenState.ScreenStatus.SUCCESS -> {
-						if (state.knownForMovies.isNotEmpty() || state.knownForSeries.isNotEmpty()) {
-							BasicText(
-								modifier = Modifier.padding(
-									start = 16.dp,
-									top = 32.dp,
-									bottom = 12.dp
-								),
-								text = stringResource(R.string.known_for),
-								style = Theme.textStyle.title.mediumMedium16
-									.copy(color = Theme.color.surfaces.onSurface)
-							)
-							LazyRow(
-								horizontalArrangement = Arrangement.spacedBy(12.dp),
-								contentPadding = PaddingValues(horizontal = 16.dp),
-                                modifier = Modifier.padding(bottom = 8.dp)
-                            ) {
-								items(state.knownForMovies) { movie ->
-									MovieCard(
-										title = movie.title,
-										vote = movie.rating,
-										imgUrl = movie.posterPath,
-										width = 124.dp,
-										aspectRatio = 0.67f,
-										modifier = Modifier.clickable { listener.onMovieClick(movie.id) }
-									)
-								}
-								items(state.knownForSeries) { series ->
-									MovieCard(
-										title = series.title,
-										vote = series.rating,
-										imgUrl = series.posterPath,
-										width = 124.dp,
-										aspectRatio = 0.67f,
-										modifier = Modifier.clickable { listener.onSeriesClick(series.id) }
-									)
-								}
-							}
-						}
-					}
-
-                    ArtistScreenState.ScreenStatus.FAILED -> {}
-                }
+                ArtistName(state)
+                ArtistDepartment(state)
+                ArtistInformation(state)
+                ArtistBiography(state)
+                ArtistMoviesAndSeries(state, listener)
             }
         }
         AppBar(
@@ -433,10 +145,4 @@ private fun ArtistScreenContent(
                 .windowInsetsPadding(WindowInsets.statusBars)
         )
     }
-}
-
-private fun formatBirthDateLegacy(birthDateLong: Long): String {
-    val date = Date(birthDateLong)
-    val formatter = SimpleDateFormat("MMMM d, yyyy", Locale.getDefault())
-    return formatter.format(date)
 }
